+++ frontend/src/ts/modals/import-export-settings.ts
@@ -1,10 +1,5 @@
-import { PartialConfigSchema } from "@monkeytype/contracts/schemas/configs";
 import * as UpdateConfig from "../config";
-import * as Notifications from "../elements/notifications";
 import AnimatedModal from "../utils/animated-modal";
-import { migrateConfig } from "../utils/config";
-import { parseWithSchema as parseJsonWithSchema } from "@monkeytype/util/json";
-import { createErrorMessage, isObject } from "../utils/misc";
 
 type State = {
   mode: "import" | "export";
@@ -48,27 +43,7 @@ const modal = new AnimatedModal({
         void modal.hide();
         return;
       }
-
-      try {
-        const parsedConfig = parseJsonWithSchema(
-          state.value,
-          PartialConfigSchema.strip(),
-          (value) => {
-            if (!isObject(value)) {
-              throw new Error("Invalid JSON");
-            }
-            return migrateConfig(value);
-          }
-        );
-        await UpdateConfig.apply(parsedConfig);
-        UpdateConfig.saveFullConfigToLocalStorage();
-        Notifications.add("Done", 1);
-      } catch (e) {
-        const msg = createErrorMessage(e, "Failed to import settings");
-        console.error(msg);
-        Notifications.add(msg, -1);
-      }
-
+      await UpdateConfig.applyFromJson(state.value);
       void modal.hide();
     });
   },
